Create eval_comparisons module
